Update function for repository class

+++ app/src/main/java/com/example/qfmenu/network/entity/CustomerDish.kt
@@ -4,5 +4,5 @@ data class CustomerDish(
     val amount: Int,
     val customerId: Long,
     val dishId: Long,
-    val promotion: Int
+    val promotion: Byte
 )
