+++ app/src/main/java/com/example/qfmenu/network/entity/Order.kt
@@ -4,6 +4,6 @@ data class Order(
     val customerId: Long,
     val nameTable: Int,
     val payments: String,
-    val promotion: Int,
+    val promotion: Byte,
     val status: String
 )
